Adding fitting Algorithm and documentation.
In Ion_model there is now a FitFreeParams class which handles the
fitting of multivariate parameters for either Yb171 or Yb174. It assumes
a set of variables and from provided x_variable_names it reduces it to
independent variables and free parameters. The `fit_func` method parses
the input data as a dictionary to be used by the `calculate_{ion}_pop`
function. The `fit` method performs the curve_fit using scipy curve_fit
function.

The user can now specify an arbitrary set of independent variables,
written as a tuple

`x = (thetaBE370, detuning935,...)`

The rest of the parameters are assumed to be free. Ideally this gets
implimented into a larger algorithm that sweeps two or more variables to
produce a multivariate curve. The fitted parameters can be used in the
next iteration where other independent variables are changed. A running
memory of the fitted parameters should help build a picture of the true
values of the free parameters. This should then lead to a mapping to the
B field vector vs the E field vector, as well as the Zeeman shift etc.

The surface of the model can be regenerated from these fitted
parameters.